auto refreshing credential
