+++ sdk/storage_datalake/Cargo.toml
@@ -28,14 +28,16 @@ serde_json = "1.0"
 serde-xml-rs = "0.5"
 uuid = { version = "0.8", features = ["v4"] }
 url = "2.2"
+tokio = { version = "1.0", optional = true }
 
 [dev-dependencies]
 tokio = { version = "1.0", features = ["full"] }
 azure_identity = { path = "../identity", version = "0.1" }
 
 [features]
-default = ["enable_reqwest", "mock_transport_framework"]
+default = ["enable_reqwest"]
 test_e2e = []
 enable_reqwest = ["azure_core/enable_reqwest"]
 enable_reqwest_rustls = ["azure_core/enable_reqwest_rustls"]
 mock_transport_framework = ["azure_core/mock_transport_framework"]
+auto_refreshing_credential = ["tokio"]
